refactor: split dashboard/runner.py into focused modules
